+++ index.html
@@ -179,7 +179,6 @@
 
             html {
                 background: #fff;
-
                 @media (prefers-color-scheme: dark) {
                     background: #000;
                 }
