+++ src/store/useUIStore.ts
@@ -1,7 +1,7 @@
 import { create } from './create';
 import { backgroundType, viewType } from './types.ts';
 import { Theme } from '@app/theme/types.ts';
-import { animationDarkBg, animationLightBg, setAnimationProperties } from '@app/visuals.ts';
+import { setAnimationProperties } from '@app/visuals.ts';
 import { useAppConfigStore } from './useAppConfigStore.ts';
 
 export const DIALOG_TYPES = ['logs', 'restart', 'autoUpdate', 'releaseNotes', 'ludicrousConfirmation'] as const;
@@ -66,3 +66,28 @@ export const setUITheme = (theme: Theme) => {
     setAnimationProperties(theme === 'light' ? animationLightBg : animationDarkBg);
     useUIStore.setState({ theme });
 };
+
+export const animationLightBg = [
+    { property: 'bgColor1', value: '#ffffff' },
+    { property: 'bgColor2', value: '#d0d0d0' },
+    { property: 'neutralColor', value: '#ffffff' },
+    { property: 'mainColor', value: '#0096ff' },
+    { property: 'successColor', value: '#00c881' },
+    { property: 'failColor', value: '#ca0101' },
+    { property: 'particlesColor', value: '#505050' },
+    { property: 'goboIntensity', value: 0.45 },
+    { property: 'particlesOpacity', value: 0.75 },
+    { property: 'particlesSize', value: 0.01 },
+];
+export const animationDarkBg = [
+    { property: 'bgColor1', value: '#212121' },
+    { property: 'bgColor2', value: '#212121' },
+    { property: 'neutralColor', value: '#040723' },
+    { property: 'successColor', value: '#c9eb00' },
+    { property: 'mainColor', value: '#813bf5' },
+    { property: 'failColor', value: '#ff5610' },
+    { property: 'particlesColor', value: '#813bf5' },
+    { property: 'goboIntensity', value: 0.75 },
+    { property: 'particlesOpacity', value: 0.95 },
+    { property: 'particlesSize', value: 0.015 },
+];
