+++ src/theme/GlobalStyle.ts
@@ -1,9 +1,8 @@
 import { createGlobalStyle } from 'styled-components';
 
 export const GlobalReset = createGlobalStyle`
-
     *:focus {
-        outline: none
+        outline: none;
     }
 
     button {
@@ -23,7 +22,7 @@ export const GlobalReset = createGlobalStyle`
             outline-offset: 2px;
         }
     }
-    
+
     fieldset,
     textarea,
     dialog,
@@ -47,7 +46,6 @@ export const GlobalReset = createGlobalStyle`
             outline-offset: 2px;
         }
     }
-
 `;
 
 export const GlobalStyle = createGlobalStyle`
@@ -67,8 +65,10 @@ export const GlobalStyle = createGlobalStyle`
         width: 100%;
         box-sizing: border-box;
         position: relative;
-        color: ${({ theme }) => theme.palette.text.primary};
-        transition: color .2s ease, background-color .2s ease,  background .2s ease;
+        transition:
+                color 0.2s ease,
+                background-color 0.2s ease,
+                background 0.2s ease;
 
         ::-webkit-scrollbar {
             display: none;
@@ -79,6 +79,11 @@ export const GlobalStyle = createGlobalStyle`
         letter-spacing: -0.02px;
         font-weight: 400;
 
+        color: #000;
+        @media (prefers-color-scheme: dark) {
+            color: #fff;
+        }
+        
         * {
             box-sizing: border-box;
 
@@ -89,14 +94,8 @@ export const GlobalStyle = createGlobalStyle`
             -ms-overflow-style: none; /* IE and Edge */
             scrollbar-width: none; /* Firefox */
         }
-
-       
     }
 
-    html {
-        background:  ${({ theme }) => theme.palette.base};
-    }
-    
     #canvas {
         z-index: 0;
         top: 0;
@@ -110,4 +109,5 @@ export const GlobalStyle = createGlobalStyle`
         z-index: 1;
         pointer-events: none;
     }
+
 `;
